Fix issu #77 - generate classes by feature creates duplicates
