+++ src/it/junit/filter-by-tag/src/test/resources/features/feature1.feature
@@ -3,3 +3,9 @@ Feature: Feature1
 
   Scenario: Matching tags are included
     Then this feature should should be included
+
+  Scenario: Another Matching tags are included
+    Then this feature should should be included
+
+  Scenario: Yet another Matching tags are included
+    Then this feature should should be included
