+++ src/it/junit/filter-by-tag/src/test/resources/features/feature2.feature
@@ -3,3 +3,9 @@ Feature: Feature2
 
   Scenario: Non Matching tags are excluded
     Then this feature should should be excluded
+
+  Scenario: More Non Matching tags are excluded
+    Then this feature should should be excluded
+
+  Scenario: And more Non Matching tags are excluded
+    Then this feature should should be excluded
