+++ src/it/junit/filter-by-tag2-negation/src/test/resources/features/feature1.feature
@@ -3,3 +3,9 @@ Feature: Feature1
 
   Scenario: Excluded tags are ignored
     Then this file will be excluded
+
+  Scenario: Another Scenario
+    Then this file will be excluded
+
+  Scenario: Yet another Scenario
+    Then this file will be excluded
