+++ src/it/junit/filter-by-tag2-negation/src/test/resources/features/feature2.feature
@@ -1,5 +1,14 @@
 @feature2
 Feature: Feature2
 
-  Scenario: No matching tags
-    Then this file will be excluded
+  Scenario: Matching tags
+    Then this file will be included
+
+  Scenario: More Matching tags
+    Then this file will be included
+
+  Scenario: Even more Matching tags
+    Then this file will be included
+
+  Scenario: Even even more Matching tags
+    Then this file will be included
