+++ src/it/junit/filter-by-tag3-and/src/test/resources/features/feature1.feature
@@ -4,3 +4,10 @@ Feature: Feature1
   @required
   Scenario: Matching tags are included
     Then this feature should should be included
+
+  @required
+  Scenario: More Matching tags are included
+    Then this feature should should be included
+
+  Scenario: Missing tags are excluded
+    Then this feature should should be excluded
