+++ src/it/junit/filter-by-tag3-and/src/test/resources/features/feature2.feature
@@ -4,3 +4,9 @@ Feature: Feature2
   @required
   Scenario: Matching tags are included
     Then this feature should should be included
+
+    
+  Scenario: Missing tags are excluded
+    Then this feature should should be excluded
+
+
