+++ src/it/junit/filter-by-tag3-and/src/test/resources/features/feature3.feature
@@ -3,3 +3,7 @@ Feature: Feature3
 
   Scenario: Non matching tags are ignored
     Then this feature should should be ignored
+    
+  Scenario: More Non matching tags are ignored
+    Then this feature should should be ignored
+
